apps/desktop/src-tauri/resources/workers/parser/requirements.txt
pytest==8.4.2
pypdf==5.9.0
pymupdf==1.25.5
python-docx==1.2.0
python-pptx==1.0.2
